using float instead of int to represent timestamp

+++ simpleutils/time.py
@@ -25,14 +25,14 @@ def to_datetime(date: Union[dt.datetime, dt.date]) -> dt.datetime:
         raise ValueError("<{0}>'s type is not recognized. Its type is <{1}>".format(date, type(date)))
 
 
-def dt2unix(time_stamp: str, tz: str = None) -> int:
+def dt2unix(time_stamp: str, tz: str = None) -> float:
     if len(time_stamp) == 25:
         date_time = pendulum.parse(time_stamp, tz=time_stamp[-6:])
     else:
         date_time = pendulum.parse(time_stamp, tz=tz)
-    return int(date_time.timestamp())
+    return date_time.timestamp()
 
 
-def unix2dt(unix_stamp: int, tz: str = None) -> str:
+def unix2dt(unix_stamp: float, tz: str = None) -> str:
     date_time = pendulum.from_timestamp(unix_stamp, tz=tz)
     return date_time.isoformat()
